bot/requirements.txt
discord
python-dotenv
psycopg2-binary